+++ src/presenters/pages/index.jsx
@@ -1,12 +1,11 @@
 import React from 'react';
 
-import {CurrentUserProvider, CurrentUserConsumer} from '../current-user.jsx';
-
-import IndexTemplate from '../../templates/pages/index';
 import LayoutPresenter from '../layout';
 import Reactlet from '../reactlet';
 
-import ByFogCreek from '../includes/by-fogcreek.jsx';
+import {getEditorUrl} from '../../models/project';
+import {CurrentUserProvider, CurrentUserConsumer} from '../current-user.jsx';
+
 import Categories from '../categories.jsx';
 import Featured from '../featured.jsx';
 import Questions from '../questions.jsx';
@@ -14,104 +13,34 @@ import RandomCategories from '../random-categories.jsx';
 import RecentProjects from '../recent-projects.jsx';
 import WhatIsGlitch from '../what-is-glitch.jsx';
 
-function old(application) {
-  console.log("Presented index");
+const FOUNDED = 2000;
+const current = new Date().getFullYear();
+const age = current - FOUNDED;
+const logo = "https://cdn.glitch.com/2bdfb3f8-05ef-4035-a06e-2043962a3a13%2Ffogcreek.svg";
+const ByFogCreek = () => (
+  <section className="by-fogcreek" role="complementary">
+    <h2>Made By Fog Creek</h2>
+    <img src={logo} alt="Fog Creek logo"/>
+    <p>
+      You might know us for making Trello, FogBugz, and co-creating Stack Overflow. 
+      We're <a href="https://www.fogcreek.com">a friendly, self-funded company</a> that's
+      been helping people make stuff for over {age} years.
+    </p>
+  </section>
+);
 
-  const self = {
-    application,
-    projects: application.projects,
+const MadeInGlitch = () => (
+  <section className="made-in-glitch">
+    <p>
+      Of course, this site was made on Glitch too{' '}
+      <a href={getEditorUrl('community')} className="button button-link has-emoji">
+        View Source <span className="emoji carp_streamer"></span>
+      </a>
+    </p>
+  </section>
+);
 
-    user: application.user,
-
-    WhatIsGlitch() {
-      if (application.currentUser().isSignedIn()) {
-        return null;
-      }
-      return Reactlet(WhatIsGlitch);
-    },
-
-    currentUser: application.currentUser,
-
-    featured() {
-      return Reactlet(Featured);
-    },
-    
-    randomCategories() {
-      const props = {
-        api: application.api(),
-      };
-      return Reactlet(RandomCategories, props);
-    },
-
-    Categories() {
-      const props = {
-        categories: application.categories,
-      };
-      return Reactlet(Categories, props);
-    },
-
-    QuestionsPresenter() {
-      const props = {
-        api: application.api(),
-      };
-      return Reactlet(Questions, props);
-    },
-
-    RecentProjectsPresenter() {
-      const props = {
-        api: application.api(),
-        userModel: application.currentUser(),
-      };
-      return Reactlet(RecentProjects, props);
-    },
-
-    ByFogCreek() {
-      return Reactlet(ByFogCreek, null);
-    },
-
-  };
-
-  const content = IndexTemplate(self);
-
-  return LayoutPresenter(application, content);
-}
-
-/*
-span
-  main(role="main")
-    h1.headline
-      a(href="https://glitch.com")
-        span= "Glitch"
-      span= " "
-      span is the friendly community where you'll build the app of your dreams
-
-    - if @currentUser().isSignedIn()
-      = @QuestionsPresenter
-    - if @currentUser().id()
-      = @RecentProjectsPresenter
-
-    = @featured
-        
-    = @randomCategories
-
-    = @Categories
-    
-    = @WhatIsGlitch
-
-    = @ByFogCreek
-
-    section.made-in-glitch
-      p
-        span Of course, this site was made on Glitch too
-      a(href="https://glitch.com/edit/#!/community")
-        button.has-emoji View Source
-          .emoji.carp_streamer
-
-  script(src="//fast.wistia.com/embed/medias/vskja9agqj.jsonp" async=true)
-  script(src="//fast.wistia.com/assets/external/E-v1.js" async=true)
-*/
-
-const IndexPage = ({api, user}) => (
+const IndexPage = ({api, categories, user}) => (
   <main>
     <h1 className="headline">
       <a href="https://glitch.com">Glitch</a>{' '}
@@ -119,6 +48,12 @@ const IndexPage = ({api, user}) => (
     </h1>
     {!!user.login && <Questions api={api}/>}
     {!!user.id && <RecentProjects api={api}/>}
+    <Featured/>
+    <RandomCategories api={api}/>
+    <Categories categories={categories}/>
+    {!user.login && <WhatIsGlitch/>}
+    <ByFogCreek/>
+    <MadeInGlitch/>
   </main>
 );
 
@@ -127,14 +62,17 @@ const IndexPageContainer = ({userModel, ...props}) => (
     <CurrentUserConsumer>
       {user => <IndexPage user={user} {...props}/>}
     </CurrentUserConsumer>
+    <script src="//fast.wistia.com/embed/medias/vskja9agqj.jsonp" async/>
+    <script src="//fast.wistia.com/assets/external/E-v1.js" async/>
   </CurrentUserProvider>
 );
 
 export default function IndexPagePresenter(application) {
   const props = {
     api: application.api(),
+    categories: application.categories,
     userModel: application.currentUser(),
   };
-  const content = Reactlet(IndexPage, props);
+  const content = Reactlet(IndexPageContainer, props);
   return LayoutPresenter(application, content);
 }
